✨ phpstan level 5

+++ Configuration/TCA/tx_maillogger_domain_model_mailtemplate.php
@@ -92,7 +92,7 @@ return [
                     'allowLanguageSynchronization' => true,
                 ],
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y')),
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y')),
                 ],
             ],
         ],
@@ -110,7 +110,7 @@ return [
                     'allowLanguageSynchronization' => true,
                 ],
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y')),
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y')),
                 ],
             ],
         ],
